+++ core-ui/src/components/App/App.js
@@ -11,172 +11,8 @@ import { ContainersLogs } from 'components/Predefined/Details/Pod/ContainersLogs
 import { ComponentForList, ComponentForDetails } from 'shared/getComponents';
 import { API_RULES_TITLE } from 'shared/constants';
 import { getResourceUrl } from 'shared/helpers';
-import jsyaml from 'js-yaml';
-
-export function ClusterList() {
-  const { clusters, currentClusterName } = useMicrofrontendContext();
-  if (!clusters) {
-    return null;
-  }
-
-  function setCluster(clusterName) {
-    LuigiClient.sendCustomMessage({
-      id: 'busola.setCluster',
-      clusterName,
-    });
-  }
-
-  function deleteCluster(clusterName) {
-    LuigiClient.sendCustomMessage({
-      id: 'busola.deleteCluster',
-      clusterName,
-    });
-  }
-
-  return (
-    <>
-      <ul>
-        {Object.entries(clusters).map(([clusterName, cluster]) => (
-          <li key={clusterName}>
-            {clusterName} {currentClusterName === clusterName && '*'}
-            <button
-              disabled={currentClusterName === clusterName}
-              onClick={() => setCluster(clusterName)}
-            >
-              set
-            </button>
-            <button onClick={() => deleteCluster(clusterName)}>delete</button>
-          </li>
-        ))}
-      </ul>
-      <button onClick={() => LuigiClient.linkManager().navigate('add')}>
-        nowy
-      </button>
-    </>
-  );
-}
-
-function readFile(file) {
-  return new Promise(resolve => {
-    const reader = new FileReader();
-    reader.onload = e => resolve(e.target.result);
-    reader.readAsText(file);
-  });
-}
-
-export function AuthForm({ auth, setAuth }) {
-  return (
-    <div>
-      <label>
-        issuer
-        <input
-          value="https://apskyxzcl.accounts400.ondemand.com"
-          onChange={e => setAuth({ ...auth, issuerUrl: e.target.value })}
-        />
-      </label>
-      <label>
-        Clientid
-        <input
-          value="d0316c58-b0fe-45cd-9960-0fea0708355a"
-          onChange={e => setAuth({ ...auth, cliendId: e.target.value })}
-        />
-      </label>
-      <label>
-        scopes
-        <input
-          value="openid"
-          onChange={e => setAuth({ ...auth, scope: e.target.value })}
-        />
-      </label>
-    </div>
-  );
-}
-
-export function AddCluster() {
-  const [showNoAuth, setShowNoAuth] = React.useState(false);
-  const [cluster, setCluster] = React.useState(null);
-  const [auth, setAuth] = React.useState({
-    issuerUrl: 'https://apskyxzcl.accounts400.ondemand.com',
-    clientId: 'd0316c58-b0fe-45cd-9960-0fea0708355a',
-    scope: 'openid',
-  });
-
-  async function onKubeconfigUploaded(file) {
-    try {
-      const kubeconfigParsed = jsyaml.load(await readFile(file));
-      handleKubeconfigAdded(kubeconfigParsed);
-    } catch (e) {
-      alert(e);
-      console.warn(e);
-    }
-  }
-
-  function handleKubeconfigAdded(kubeconfig) {
-    const clusterName = kubeconfig.clusters[0].name;
-    try {
-      const cluster = {
-        name: clusterName,
-        server: kubeconfig.clusters[0].cluster.server,
-        'certificate-authority-data':
-          kubeconfig.clusters[0].cluster['certificate-authority-data'],
-      };
-      const user = kubeconfig.users[0].user;
-      const token = user.token;
-      const clientCA = user['client-certificate-data'];
-      const clientKeyData = user['client-key-data'];
-      if (token || (clientCA && clientKeyData)) {
-        const params = {
-          cluster,
-          rawAuth: {
-            idToken: token,
-            'client-certificate-data': clientCA,
-            'client-key-data': clientKeyData,
-          },
-        };
-        LuigiClient.sendCustomMessage({
-          id: 'busola.addCluster',
-          params,
-        });
-      } else {
-        setShowNoAuth(true);
-        setCluster(cluster);
-      }
-    } catch (e) {
-      alert(e);
-      console.warn(e);
-    }
-  }
-
-  function dupa() {
-    const params = {
-      cluster,
-      auth: { ...auth, responseType: 'id_token' },
-    };
-    console.log(params);
-    LuigiClient.sendCustomMessage({
-      id: 'busola.addCluster',
-      params,
-    });
-  }
-
-  return (
-    <>
-      <button onClick={() => LuigiClient.linkManager().navigate('/clusters')}>
-        back
-      </button>
-      <input
-        type="file"
-        onChange={e => onKubeconfigUploaded(e.target.files[0])}
-      />
-      {showNoAuth && (
-        <>
-          <AuthForm auth={auth} setAuth={setAuth} />
-          <button onClick={dupa}>dawaj</button>
-        </>
-      )}
-    </>
-  );
-}
+import { ClusterList } from 'components/Clusters/ClusterList';
+import { AddCluster } from 'components/Clusters/AddCluster';
 
 export default function App() {
   return (
