+++ core-ui/src/components/Clusters/shared.js
@@ -1,0 +1,30 @@
+import LuigiClient from '@luigi-project/client';
+
+export function setCluster(clusterName) {
+  LuigiClient.sendCustomMessage({
+    id: 'busola.setCluster',
+    clusterName,
+  });
+}
+
+export function addCluster(params) {
+  LuigiClient.sendCustomMessage({
+    id: 'busola.addCluster',
+    params,
+  });
+}
+
+export function deleteCluster(clusterName) {
+  LuigiClient.sendCustomMessage({
+    id: 'busola.deleteCluster',
+    clusterName,
+  });
+}
+
+export function readFile(file) {
+  return new Promise(resolve => {
+    const reader = new FileReader();
+    reader.onload = e => resolve(e.target.result);
+    reader.readAsText(file);
+  });
+}
