+++ core/src/luigi-config/auth-storage.js
@@ -1,0 +1,13 @@
+let authData;
+
+export function clearAuthData() {
+  authData = null;
+}
+
+export function setAuthData(data) {
+  authData = data;
+}
+
+export function getAuthData() {
+  return authData;
+}
