+++ core/src/luigi-config/navigation/static-navigation-model.js
@@ -647,6 +647,7 @@ export function getStaticRootNodes(namespaceChildrenNodesResolver, apiGroups) {
         }),
       icon: 'dimension',
       viewGroup: coreUIViewGroupName,
+      navigationContext: 'namespaces',
     },
     {
       pathSegment: 'namespaces',
