+++ core/src/luigi-config/settings.js
@@ -4,7 +4,7 @@ export function createSettings(params) {
     sideNavFooterText: '',
     header: {
       logo: 'assets/logo.svg',
-      title: getTitle(params),
+      title: params ? params.cluster.name : 'Busola',
       favicon: 'favicon.ico',
     },
     appLoadingIndicator: {
@@ -12,13 +12,4 @@ export function createSettings(params) {
     },
     customSandboxRules: ['allow-downloads'],
   };
-}
-
-function getTitle(params) {
-  if (!params) {
-    return 'Busola';
-  } else {
-    const cluster = params.cluster;
-    return cluster.name || cluster.server.replace(/^https?:\/\/(api\.)?/, '');
-  }
 }
